Ensure converted timestamp is an int
DynamoDB doesn't like floats, and we want an int anyway. Python tries to
be "helpful" and lets int/int=float if needed, which is nice in a math
sence, terrible in many use cases, though.

+++ src/F2B/__init__.py
@@ -22,7 +22,8 @@ def process_log_events(log_events, parsers, source=None):
                 match = {
                     'Host': resp['host'],
                     'MatchID': uuid5(parser_instance.uuid, event['id']),
-                    'Timestamp': event['timestamp']/1000  # Convert miliseconds
+                    # Convert miliseconds
+                    'Timestamp': int(event['timestamp']/1000)
                 }
                 log.debug('Match data: {0}'.format(match))
                 add_match(match['Host'], match['MatchID'], parser_name,
